% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchAdjust.R
\name{batchAdjust}
\alias{batchAdjust}
\title{batchAdjust}
\usage{
batchAdjust(envData, methylData, batchName = FALSE)
}
\arguments{
\item{envData}{dataframe holding all environmental/phenotype data}

\item{methylData}{dataframe holding all methylation data}

\item{batchName}{string variable. If equal to FALSE then no batch effect adjustment should be made, otherwise string is set to the name of the variable holding the batch labels.}
}
\value{
output returns dataframe of methylation data
}
\description{
This function takes two files as input, the methylation data and the environment/batch data
 and uses the ComBat method to adjust for the batch effect
}
