% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covar.R
\docType{data}
\name{covar}
\alias{covar}
\title{Simulated covariaten data set with 11 variables.}
\format{A data frame with 987 rows and 11 columns:
\describe{
  \item{ID}{An ID column, numers from 1 to 987}
  \item{sex}{A variable representing biological sex, coded as 0 and 1}
  \item{BMI}{A randomly generated body mass index (BMI) variable}
  \item{preg_smoking}{A randomly generated variable for exposure to maternal smoking, coded as 0 and 1}
  \item{Gran}{Randomly generated granulocyte cell counts}
  \item{Mono}{Randomly generated monocytes cell counts}
  \item{Bcell}{Randomly generated beta cells cell counts}
  \item{NK}{Randomly generated natural killer cells cell counts}
  \item{CD4T}{Randomly generated CD4T cell counts}
  \item{CD8T}{Randomly generated CD8T cell counts}
  \item{Plate_no}{Randomly generated batch variable}
}}
\usage{
covar
}
\description{
A dataset containing 11 simulated variables and 987 rows
}
\keyword{datasets}
