% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth.R
\docType{data}
\name{meth}
\alias{meth}
\title{Simulated methylation data set with 1000 CpGs.}
\format{A data frame with 1000 rows and 988 columns:
\describe{
  \item{cgXXXXXX}{DNA methylation beta}
  ...
}}
\usage{
meth
}
\description{
A dataset containing a simulated subset of 1000 CpGs from the Illumina HumanMethylation450 array
}
\keyword{datasets}
