% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyGEM.R
\name{shinyGEM}
\alias{shinyGEM}
\title{shinyGEM}
\usage{
shinyGEM()
}
\value{
\code{shinyGEM} to launch shiny app
}
\description{
Shiny app to make the GEM model approach accessible to everyone
}
\examples{
\dontrun{
 shinyGEM()
 }
}
