% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyGEM_Emodel.R
\name{shinyGEM_Emodel}
\alias{shinyGEM_Emodel}
\title{The GEM_Emodel method is used to find associations between methylation and environmental factors on a epigenome wide level.}
\usage{
shinyGEM_Emodel(envFileName, methylFileName, predictorName,
  batchName = FALSE, covName = NULL, qqplotInclude = TRUE)
}
\arguments{
\item{envFileName}{A csv or txt file name of demographic and phenotypical data with columns representing environmental factors and columns representing sample}

\item{methylFileName}{A csv or txt file of methylation data with the CpG sites in the rows and the samples across the columns}

\item{predictorName}{= "Smoke" a variable continaing the name of the target environmental factor}

\item{batchName}{= "Plate_No" a variable containing the name of the batch effect variable used for the ComBat methylation}

\item{covName}{Either a single string or a vector of string containing the column names included in the regression model}

\item{output_file_name}{= "GemEmodelOutput" a file name used as the output of the funciton. All results will be written to a file with this name.}
}
\value{
output a list of values for the
}
\description{
GEM_Emodel adjusts the methylation data for batch effect, if required, using the sva package. It finds associations between
methylation and environment by performing matrix
based iterative correlation and memory-efficient data analysis instead of millions of linear regressions
(N = number_of_CpGs). The methylation data are the measurements for CpG probes, for example, 450,000 CpGs
from Illumina Infinium HumanMethylation450 Array. The environmental factor can be a particular phenotype or environment
factor from,for example, birth outcomes, maternal conditions or disease traits. The output of GEM_Emodel
for particular environmental factor is a list of CpGs that are potential epigenetic biomarkers.
GEM_Emodel runs linear regression like lm (M ~ E + covt), where M is a matrix with methylation data,
E is a matrix with environment factor and covt is a matrix with covariates, and all read from the
formatted text data file.
}
