% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VolcanoPlot.R
\name{VolcanoPlot}
\alias{VolcanoPlot}
\title{Create a Volcano plot}
\usage{
VolcanoPlot(EWAS, title = "title")
}
\arguments{
\item{EWAS}{An EWAS results data frame with columns for CpG-ID, p-value,}

\item{title}{Specify a title for the Volcano plot}
}
\value{
\code{VolcanoPlot} as an image.
}
\description{
This function utilizes the \code{annotateCpG()} function of this package. It
creates a Volcano plot with annotated gene names based on the \code{IlluminaHumanMethylation450kanno.ilmn12.hg19} bioconductor package.
}
\details{
IMPORTANT: The CpG identifier column needs to be called "ID"!
}
